delete_pruebasUnits

+++ src/main/java/co/edu/eam/ingesoft/stores/services/StoresService.java
@@ -1,18 +1,16 @@
 package co.edu.eam.ingesoft.stores.services;
-
 import java.util.List;
+import javax.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import co.edu.eam.ingesoft.stores.model.Stores;
 import co.edu.eam.ingesoft.stores.repositories.StoresRepository;
-
 /**
  * Bussiness class for Stores Entity.
  *
  * @author caferrerb
  *
  */
-
 @Service
 public class StoresService {
   /**
@@ -45,7 +43,6 @@ public class StoresService {
   public void update(Stores store) {
     storesRepository.save(store);
   }
-
   /**
    * List all stores.
    *
@@ -54,16 +51,20 @@ public class StoresService {
   public List<Stores> listAll() {
     return storesRepository.findAll();
   }
-
   /**
-   * Method delete a store.
+   * Delete a Store.
    *
-   * @param id to delete
+   * @param id id to delete
+   * @return store deleted
    */
-  public void delete(String id) {
+  public Stores delete(String id) {
+    Stores stores = find(id);
+    if (stores == null) {
+      throw new EntityNotFoundException("Store not exists");
+    }
     storesRepository.deleteById(id);
+    return stores;
   }
-
   /**
    * List stores by name.
    *
